JBIDE-18041 Visual Editor: refresh type preference is not saved
fixed handling of norefresh option
