+++ plugins/org.jboss.tools.vpe.preview.core/src/org/jboss/tools/vpe/preview/core/util/ActionBar.java
@@ -1,3 +1,13 @@
+/*******************************************************************************
+ * Copyright (c) 2014 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v1.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributor:
+ *     Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
 package org.jboss.tools.vpe.preview.core.util;
 
 import java.net.MalformedURLException;
@@ -19,9 +29,14 @@ import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.commands.ICommandService;
 import org.jboss.tools.vpe.preview.core.Activator;
 
-public class ActionBar {
+/**
+ * @author Konstantin Marmalyukov (kmarmaliykov)
+ */
+
+public abstract class ActionBar {
 	private static final String GROUP_REFRESH = "org.jboss.tools.vpv.refresh"; //$NON-NLS-1$
 	private static final String REFRESH_ON_SAVE_ENABLEMENT = "org.jboss.tools.vpe.enableRefreshOnSave"; //$NON-NLS-1$
+	private static final String REFRESH_ON_CHANGE_ENABLEMENT = "org.jboss.tools.vpe.enableRefreshOnChange"; //$NON-NLS-1$
 
 	private IAction refreshAction;
 	private IAction openInDefaultBrowserAction;
@@ -29,7 +44,6 @@ public class ActionBar {
 	private IAction enableRefreshOnSaveAction;
 	
 	private IExecutionListener saveListener;
-	private boolean enableAutomaticRefresh;
 	private Browser browser;
 	
 	private IPreferenceStore preferences;
@@ -40,7 +54,6 @@ public class ActionBar {
 	public ActionBar(Browser browser1, IPreferenceStore preferences) {
 		this.browser = browser1;
 		this.preferences = preferences;
-		enableAutomaticRefresh = !preferences.getBoolean(REFRESH_ON_SAVE_ENABLEMENT);
 		ICommandService commandService = (ICommandService) PlatformUI.getWorkbench().getService(ICommandService.class);
 		saveCommand = commandService.getCommand("org.eclipse.ui.file.save"); //$NON-NLS-1$
 		saveAllCommand = commandService.getCommand("org.eclipse.ui.file.saveAll"); //$NON-NLS-1$
@@ -87,32 +100,21 @@ public class ActionBar {
 		enableRefreshOnSaveAction.run();
 	}
 	
-	protected void refresh(Browser browser) {
-		String url = browser.getUrl();
-		if (PlatformUtil.isWindows()) {
-			url = NavigationUtil.removeAnchor(url);
-		}
-		browser.setUrl(url);
-	}
+	protected abstract void refresh(Browser browser);
 
 	private void makeEnableAutomaticRefreshAction() {
 		enableAutomaticRefreshAction = new Action(Messages.VpvView_ENABLE_AUTOMATIC_REFRESH, IAction.AS_CHECK_BOX) {
 			@Override
 			public void run() {
 				if (enableAutomaticRefreshAction.isChecked()) {
-					enableAutomaticRefresh = true;
-
-					preferences.setValue(REFRESH_ON_SAVE_ENABLEMENT, false);
 					enableRefreshOnSaveAction.setChecked(false);
 					saveCommand.removeExecutionListener(saveListener);
 					saveAllCommand.removeExecutionListener(saveListener);
-				} else {
-					enableAutomaticRefresh = false;
 				}
 			}
 		};
 
-		enableAutomaticRefreshAction.setChecked(!preferences.getBoolean(REFRESH_ON_SAVE_ENABLEMENT));
+		enableAutomaticRefreshAction.setChecked(preferences.getBoolean(REFRESH_ON_CHANGE_ENABLEMENT));
 		enableAutomaticRefreshAction.setImageDescriptor(Activator.getImageDescriptor("icons/refresh_on_change.png")); //$NON-NLS-1$
 	}
 
@@ -123,10 +125,8 @@ public class ActionBar {
 				if (enableRefreshOnSaveAction.isChecked()) {
 					saveCommand.addExecutionListener(saveListener);
 					saveAllCommand.addExecutionListener(saveListener);
-					
-					preferences.setValue(REFRESH_ON_SAVE_ENABLEMENT, true);
+
 					enableAutomaticRefreshAction.setChecked(false);
-					enableAutomaticRefresh = false;
 				} else {
 					saveCommand.removeExecutionListener(saveListener);
 					saveAllCommand.removeExecutionListener(saveListener);
@@ -169,11 +169,14 @@ public class ActionBar {
 	
 	
 	public boolean isAutomaticRefreshEnabled() {
-		return enableAutomaticRefresh;
+		return enableAutomaticRefreshAction.isChecked();
 	}
 	
 	public void dispose() {
 		saveCommand.removeExecutionListener(saveListener);
 		saveAllCommand.removeExecutionListener(saveListener);
+		
+		preferences.setValue(REFRESH_ON_CHANGE_ENABLEMENT, enableAutomaticRefreshAction.isChecked());
+		preferences.setValue(REFRESH_ON_SAVE_ENABLEMENT, enableRefreshOnSaveAction.isChecked());
 	}
 }
