+++ plugins/org.jboss.tools.vpe.preview.editor/src/org/jboss/tools/vpe/preview/editor/VpvEditor.java
@@ -318,7 +318,7 @@ public class VpvEditor extends EditorPart implements VpvVisualModelHolder, IReus
 	}
 	
 	public void addPreviewToolbarItems() {
-		actionBar = new ActionBar(browser, WebUiPlugin.getDefault().getPreferenceStore());
+		actionBar = new VisualEditorActionBar(browser, WebUiPlugin.getDefault().getPreferenceStore(), controller);
 		actionBar.fillLocalToolBar(toolBarManager);
 		toolBarManager.update(true);
 	}
